refactor activity page
